requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning
torchmetrics

nltk
datasets
triton>=2.2.0
einops
lmdb
datasets
scikit-learn
nltk
pytrec_eval
sklearn-contrib-lightning
pandas
transformers
datasets